US O&G Wells
Oil and gas wells in the US February 2014.  This excludes data from TX,
MD, and NC, which could not be obtained when this file was created, and
wells in OK and CA are significantly under-represented.  For better data
on those states, please contact the respective oil and gas regulatory
offices.